вынес в конфигурационный класс
