+++ src/main/java/faang/school/postservice/service/PostService.java
@@ -1,7 +1,9 @@
 package faang.school.postservice.service;
 
+import faang.school.postservice.client.LanguageToolClient;
 import faang.school.postservice.dto.post.PostRequestDto;
 import faang.school.postservice.dto.post.PostResponseDto;
+import faang.school.postservice.exception.LanguageToolException;
 import faang.school.postservice.exception.PostNotFoundException;
 import faang.school.postservice.mapper.PostMapper;
 import faang.school.postservice.model.Post;
@@ -9,11 +11,26 @@ import faang.school.postservice.repository.PostRepository;
 import faang.school.postservice.utils.validationUtils.PostValidation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.retry.annotation.Backoff;
+import org.springframework.retry.annotation.Recover;
+import org.springframework.retry.annotation.Retryable;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 @Service
 @Slf4j
@@ -22,9 +39,17 @@ public class PostService {
     public static final String CANT_UPDATE_DELETED_POST = "Can't update deleted post";
     public static final String NO_POST_FOUND = "No post found with ID %d";
     public static final String POST_HAS_ALREADY_BEEN_DELETED = "Post has already been deleted";
+    private static final int TIMEOUT_HOURS = 2;
 
     private final PostMapper postMapper;
     private final PostRepository postRepository;
+    private final LanguageToolClient languageToolClient;
+
+    @Value("${posts.correction.batch-size}")
+    int batchSize;
+
+    @Value("${posts.correction.thread-poop-size}")
+    int threadPoolSize;
 
     public PostResponseDto createDraftPost(PostRequestDto postRequestDto) {
         PostValidation.validatePostAuthors(postRequestDto);
@@ -114,5 +139,63 @@ public class PostService {
             log.error(message);
             throw new PostNotFoundException(message);
         }
+    }
+
+    public void sendPostsForChecking() {
+        ExecutorService executor = Executors.newFixedThreadPool(threadPoolSize);
+        long total = postRepository.count();
+        List<CompletableFuture<Void>> futures = new ArrayList<>();
+        AtomicBoolean hasErrors = new AtomicBoolean(false);
+
+        for (int start = 0; start < total; start += batchSize) {
+            int end = Math.min(start + batchSize - 1, (int) total);
+            int size = end - start + 1;
+            int finalStart = start;
+            futures.add(CompletableFuture.runAsync(() -> {
+                Pageable pageable = PageRequest.of((finalStart + size - 1) / size, size);
+                Page<Post> postContents = postRepository.findPosts(pageable);
+                sendPostContentsChecking(postContents.getContent(), hasErrors);
+            }, executor));
+        }
+
+        try {
+            CompletableFuture.allOf(futures.toArray(new CompletableFuture[0])).get(TIMEOUT_HOURS, TimeUnit.HOURS);
+            if (hasErrors.get()) {
+                log.warn("Not all posts have been corrected");
+            } else {
+                log.info("Submitting posts for review completed successfully");
+            }
+        } catch (TimeoutException e) {
+            log.error("Submitting posts for review haven't completed on time");
+        } catch (InterruptedException e) {
+            log.error("Submitting posts for review was interrupted. {}", e.getMessage());
+            Thread.currentThread().interrupt();
+        } catch (ExecutionException e) {
+            log.error("Execution exception while submitting posts for review. {}", e.getMessage());
+        }
+    }
+
+    @Retryable(
+            retryFor = {LanguageToolException.class},
+            maxAttemptsExpression = "${spring.retry.language-tool.max-attempts}",
+            backoff = @Backoff(delayExpression = "${spring.retry.language-tool.backoff-delay}")
+    )
+    private void sendPostContentsChecking(List<Post> posts, AtomicBoolean hasErrors) {
+        for (Post post : posts) {
+            log.debug("Before correcting errors in the text: {}", post.getContent());
+            try {
+                post.setContent(languageToolClient.getCorrectedText(post.getContent(), "auto").block());
+                postRepository.save(post);
+                log.debug("After correcting errors in the text: {}", post.getContent());
+            } catch (LanguageToolException e) {
+                hasErrors.set(true);
+                throw e;
+            }
+        }
+    }
+
+    @Recover
+    private void recoverSendPostContentsChecking(LanguageToolException e) {
+        log.error("Failed to correct text after retries. {}", e.getMessage());
     }
 }
